requirements.txt
python=3.6.10
pytorch=1.6.0
numpy=1.18.4
tqdm=4.46.0
